requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib
bcrypt==3.2.2
PyJWT
email-validator